callback to separate module
