+++ account-abstraction/src/lib.rs
@@ -3,6 +3,7 @@
 #[allow(unused_imports)]
 use multiversx_sc::imports::*;
 
+pub mod custom_callbacks;
 pub mod signature;
 pub mod unique_payments;
 pub mod user_actions;
@@ -13,6 +14,7 @@ pub trait AccountAbstraction:
     users::UsersModule
     + signature::SignatureModule
     + user_actions::execution::ExecutionModule
+    + custom_callbacks::CustomCallbacksModule
     + utils::UtilsModule
 {
     #[init]
