+++ account-abstraction/src/user_actions/execution.rs
@@ -1,4 +1,5 @@
 use crate::{
+    custom_callbacks::CallbackProxy as _,
     signature::CheckExecutionSignatureArgs,
     unique_payments::{PaymentsVec, UniquePayments},
 };
@@ -13,7 +14,10 @@ multiversx_sc::imports!();
 
 #[multiversx_sc::module]
 pub trait ExecutionModule:
-    crate::users::UsersModule + crate::signature::SignatureModule + utils::UtilsModule
+    crate::users::UsersModule
+    + crate::signature::SignatureModule
+    + crate::custom_callbacks::CustomCallbacksModule
+    + utils::UtilsModule
 {
     #[endpoint(multiActionForUser)]
     fn multi_action_for_user(
@@ -152,24 +156,5 @@ pub trait ExecutionModule:
 
     fn require_non_empty_actions<T>(&self, actions: &MultiValueEncoded<T>) {
         require!(!actions.is_empty(), "No actions");
-    }
-
-    #[callback]
-    fn user_action_cb(
-        &self,
-        original_caller: ManagedAddress,
-        original_payments: PaymentsVec<Self::Api>,
-        #[call_result] call_result: ManagedAsyncCallResult<IgnoreValue>,
-    ) {
-        if call_result.is_ok() {
-            return;
-        }
-
-        let user_id = self.user_ids().get_id(&original_caller);
-        self.user_tokens(user_id).update(|user_tokens| {
-            for payment in &original_payments {
-                user_tokens.add_payment(payment);
-            }
-        });
     }
 }
